remove unneccesary optimizations
