+++ ydb/core/blobstorage/vdisk/common/vdisk_costmodel.h
@@ -2,7 +2,6 @@
 
 #include "defs.h"
 #include "vdisk_events.h"
-#include <ydb/core/base/blobstorage.h>
 
 
 namespace NKikimr {
